Clean up rebinding visitors.
